display landing page only in demo mode
